[Feat] #11-Enable VoteButton when OptionCVC Clicked
